Tweak of tagline
This is following feedback that it could be read as exclusionary to people who have other kinds of technical knowledge. New wording to make it clearer that you don’t NEED to have any technical knowledge to use the platform.

Also changed 'technical' to 'coding' in the 'Easy to use' section to be more specific, and to avoid repeating 'technical'.