Add jest and puppeteer test config
To support new client side javascript features, cookie consent and
analytics, this commit adds Jest and puppeteer to test code.

Currently there are only example tests to ensure that the configuration
is correct. More tests will be added in later commits.

Two separate 'projects', for tests are created in the jest
config.

The first is for unit tests which are run using jsdom. Files under
/source/javascripts with a filename ending .test.mjs are run when `yarn
run test:unit` is used. Jsdom is used for unit tests because without it
code has to be written without relying on the standard browser
interface, which is possible but tends to make code more convoluted.
Jsdom is preferred over running these tests using puppeteer as its a lot
quicker.

The second set of tests, found in /tests/functional are run using
puppeteer, which starts a middleman server and runs tests using a
headless browser. These test bundled javascript code along with HTML and
should tests functionality more.

Running `yarn run test` runs both tests, but not the linter, which
should be run separately.