Set consent-cookie path
Cookies are set with a path value which defaults to the current page.
This meant that when consent was granted or denied on different pages,
the cookie wasn't read on others.

This is fixed by setting the path to '/'. This is ok because the site is
hosted on 'https://www.forms.service.gov.uk/'