Updates to features page.html.erb
Added brief info about detailed guidance to 'Features available now' and removed it from 'Features we're working on next'.

Updated 'Features we're working on next' to include a metrics dashboard and confirmation email when someone's submitted a form. Moved tabular questions to the bottom of this list to reflect changing road map priorities.