Use setup-node action to install Node.js and cache npm
Currently `npm ci` can take over 2 minutes in GitHub Actions, this
commit uses the actions/setup-node action to cache the npm downloads to
try and speed things up.

We also use the action to install the same version of Node as in the
`.nvmrc` file, because we might as well.