Change default port
To avoid collisions with the forms-admin app [[1]], change the default
port from 3000 to 3002.

Note that for some reason, this change doesn't take effect when running
the app in a Docker container, so we don't change the Dockerfile or any
deployment associated code. We're not sure why this is happening, so we
might revisit, but for now the behaviour is consistent, with both this
app and forms-runner [[2]], so until we understand what is going on
better we'll leave things as is.

[1]: https://github.com/alphagov/forms-admin/blob/545be0398d190c29b4bacbec6ac4955864595b23/config/puma.rb#L18
[2]: https://github.com/alphagov/forms-runner/blob/2f235b04ba8c41156a1523f48c1840063e82ee49/config/puma.rb#L18