Replace header HTML with view component
Use the header component instead of putting the HTML straight into the
layout templace.

This commit also adds a regression spec for the header; we want to make
sure any customizations we applied to the hardcoded HTML get carried
over to our component. This commit moves the existing HTML into the
header component spec file, and adds a test that compares the output of
the rendered component with that, modulo some normalization of
formatting. Unfortunately the normalization isn't perfect, so the test
is currently failing, and is marked as such.