Slime regression test
Try and make the header regression test pass by correcting a few small
non-functional differences, expanding some ERB, and also removing the `id`
of the `main` tag which we never used.

Unfortunately it turns out this regression test works a little too well,
and it spotted a few bugs in the upstream govuk_components gem 😅. See
issue x-govuk/govuk-components#459 for details. So for now this test is
still failing.