Add workaround for incorrect header ARIA labels
The `govuk_components` gem has outdated default content for the header;
the ARIA labels were changed in GOV.UK Frontend v4 to improve the
announcements from screen readers, but that change didn't make it into
the gem (which is managed by a different team).

This commit updates our configuration of the Govuk::Components module so
that the default for our app matches what is in the latest release of
govuk-frontend [[1]].

See x-govuk/govuk-components#460 for more details.

[1]: https://github.com/alphagov/govuk-frontend/blob/v4.7.0/package/govuk/components/header/template.njk#L65-L66