Update features page
Removed a line about us working on trial accounts as we've done this now. Not including this elsewhere in this page as we normally would as it's not a feature people need to be aware of for deciding if this meets their needs. 

Have added "Taking a payment with a completed form" as that's next on the roadmap along with 'add another answer'.