Remove logging of the Zendesk response body
If there was an error during creation of a Zendesk ticket via the API,
we were raising an exception that contained the whole Zendesk response
body. This has resulted in two incidents where PII was logged to
Sentry, so we are just removing this logging entirely for now. We could
look at more nuanced scrubbing of the data in future.