Update icon paths
govuk-frontend v5 has a slightly different set of icons compared to v4,
so we need to update the icon paths we reference in
application.html.erb.

We're adding the new manifest and icons to the public folder, because
fingerprinting the images breaks the paths in the manifest. We should
find a way to either generate the manifest with references to
fingerprinted assets, or automatically copy the manifest and its
dependencies to `public`.