Update Features page content
Removed references to getting a notification when GOV⁠.⁠UK Forms has launched - as it has now. 

Removed the link to the mailing list as it’s at the bottom of the page now.

And removed references to it being in development with selected government partners.

More improvements to this page to follow.