Extract core dependency versions into build args
This commit adds build arguments [[1]] which define the base image to
use for our app image.

These build arguments are not intended to be set by `docker build`
command line arguments, instead we just want to use the default values
for interpolation later in the Dockerfile. This makes it easier for the
`update_app_versions script` to update the versions of Alpine Linux,
Ruby and Node.js in the Dockerfile (see alphagov/forms-deploy#503).

We have to a bit of a dance with the `NODEJS_VERSION` arg because we
want to use it in the context of the `BUILD` layer, but ARG variables go
out of scope at the end of each build stage [[2]]. ENV variables are
inheritied by stages [[3]], but we can't set ENV variables outside of a
build context. So we have to use ARG variables to define the version of
Ruby and Alpine Linux, and we can't set the version of Node.js with the
other build args.

[1]: https://docs.docker.com/engine/reference/builder/#arg
[2]: https://docs.docker.com/engine/reference/builder/#scope
[3]: https://docs.docker.com/engine/reference/builder/#env