Update govuk-components to include tudor crown
Govuk-components, the ruby library we use to consume the design system
has been updated to include the new tudor crown icon in the header.

There has not been a release yet though.

This commit changes the reference in the Gemfile to use a git commit to
the main branch of the library, rather than the offical ruby gem with
version.

This allows us to release the new tudor crown changes and update the
Gemfile when the library makes a release.

Because the header is changing, we must also update the header component
tests to reflect the changes to the SVG.