Fix feature tests which use choose on radios
Feature tests for the support form fail with the following error about
selecting radio buttons:
```
Selenium::WebDriver::Error::ElementClickInterceptedError:
       element click intercepted: Element <input id="support-form-i-need-help-with-other-government-service-field" class="govuk-radios__input" type="radio" value="other_government_service" name="support_form[i_need_help_with]"> is not clickable at point (134, 656). Other element would receive the click: <label for="s
upport-form-i-need-help-with-other-government-service-field" class="govuk-label govuk-radios__label">...</label>
         (Session info: chrome=121.0.6167.184)
```

For example: https://github.com/alphagov/forms-product-page/actions/runs/7902388230/job/21567958989?pr=231

It's not clear why this issue is showing up now.

The solution is to allow clicking on labels when selecting radio buttons
(which is what the user does given the govuk styling).

In the Admin we solved this in this commit configuring capybara:

https://github.com/alphagov/forms-admin/commit/be1d26267ee4b222cbedb512f67c65eba8752f2e#diff-8c30e7862b0fec0ff5a6f5d1faefbaa8d9b6049f581f1290a2d83017ae072295R41

This commit uses the same technique.