Rake tasks for pushing commits to the dev environment
To push a commit to a dev environment, we invoke an image building pipeline to
produce an image with a specific prefix ("dev-"), which triggers the dev
deployment pipelines.

This commit adds two Rake tasks:
* push_to_dev
* reset_dev_to_main

The former deploys the commit at the tip of the current branch to dev. The
latter deploys the commit at the tip of main, after updatig the remotes to get
the very latest commits.