Ignore rubocop rule to fix broken test
Changing this to use the class constant directly caused us to get
`uninitialized constant Event` errors when running the test. This commit
reverts the change and disables the cop on this line to fix the tests.

We should try and re-enable this cop in future but I think this should
be acceptable for getting the dependabot update merged.