Update rubocop-govuk to rubocop-rails-omakase
There are far too many overrides in rubocop-govuk that has made the linter too lenient on allowing non-idiomatic code to be written by devs who are newer to Ruby on Rails. A lot of conventions are not learned. This is an exercise to get that into practise and then swap back. We won't get rid of all the overrides, but will slowly reintroduce the useful ones.