Remove devDependency for esbuild
The bundler esbuild is listed in package.json as a dev dependency.

We use esbuild through vite, which itself includes it as a dependency.

I don't think we need to specify it explicitly in our own package.json
file.

Removing it means less dependency bot updates. This matches the other
projects which don't list it.