Tell Dependabot to manage vite_ruby
vite_ruby is a transitive dependency of vite_rails in the Bundler
ecosystem, and also a implicit dependency of vite-plugin-ruby in the npm
ecosystem.

We don't want to get Dependabot PRs for all our indirect dependencies,
but we do want to make sure vite_ruby is kept up-to-date, so this commit
updates the configuration to explicitly tell Dependabot to manage that
gem.

We also add a line telling Dependabot to manage all direct dependencies
in Bundler, to make sure the default behaviour is still there.