Load govuk-frontend Sass with NodePackageImporter
There's a bug in the way Vite's inbuilt sass importing works that means
some of our assets don't build properly if we import the govuk-frontend
styles using the @govuk alias.

We can get around this by using Sass's builtin NodEpackageImporter to
import the govuk-frontend styles. This bypasses the Vite import code
that causes the issue.

It's also better for a couple of other reasons. It makes it clear that
this import is coming from an npm package (which was obscured slightly
when we were using the alias). It also makes us resilient to structure
changes in the govuk-frontend package (NodePackageImporter uses the
'sass' field in govuk-frontend's package.json to determine the file to
import, so we no longer need to point to a specific file which could
change).