Workaround for capybara failure
We're seeing an issue where it appears that Capybara isn't waiting for
the page navigation to complete before checking the URL, resulting in a
wrongly failing test.

For some matchers, such as have_text, Capybara waits for the text to
become available. That wait logic still appears to be working correctly.

This commit switches the order of our have_text and current_url
assertions, so that the current_url assertion benefits from have_text
matcher's wait logic.