Simplify header component
Removes the navigation links from the header, since they're now included
in the service navigation component.

Also simplifies the tests, in particular removing one that was very
prescriptive about the HTML structure. I believe this has caused us
issues in the past (with minor govuk-frontend HTML changes breaking our
test) so I think it makes sense to only test the things we know we need
the component to display.