Style sign in link separately
We want the sign in link to appear at the right of the service
navigation to match the design.

To make this easy, this commit adds a `featured_link` argument to the
service navigation component, which takes a hash in the same format as
the other navigation items. The component will then add the necessary
CSS classes to style it at the right of the component.