Add volume hints to the Dockerfile
We can use VOLUME instructions in our Dockerfile to instruct the runtime to
create a mount point for an external volume.

When AWS ECS sees the instruction (from the built container), it automatically
attaches ephemeral storage at that location [1]. This is useful for us because
we want to enable read only root filesystems, but the applications still need
to be able to write to a small portion of it. We can achieve that by mounting
ephemeral storage atop the paths it needs to write to.

[1] https://docs.aws.amazon.com/AmazonECS/latest/developerguide/bind-mounts.html#bind-mount-considerations