Create volumes for whole directories in the Dockerfile
Previously we created 1 volume for each directory that the application needed
to write to, because we found it didn't have permission to create
subdirectories of a volume.

This is because we weren't doing it correctly. The correct way to do it is to
create the directory in the Dockerfile and set its permissions and owner, then
expose it as a volume.