Remove Beta phase tag from header
After the rebrand, the phase tag causes the product name to reflow to
the next line on mobile displays with a small viewport (320px).
Notify did some research on this and found that it didn't mean much to
users so decided to remove it when they were in beta, so we're
following suit.