Add class to cookies tables to make text smaller on mobile
The new type scale causes the font size to be the same on mobile
displays as on desktop. For tables with lots of information, this
means we are able to fit less information in them without. Use the
govuk-table--small-text-until-tablet class so that the text is smaller
in tables on mobile devices.