Move vite assets to /assets
Currently the vite build outputs to public/vite/assets. This doesn't
match the runner and admin configuration which outputs them to
public/assets.

This doesn't matter on it's own but it's a bit confusing to have
them in different places.

If we wanted to apply caching at the CDN level, it's easier if they are
in a standard location.

This commit aligns the configuration with the runner and admin by
putting them in /public/assets.

There shouldn't be any changes to build or how the site works.