Add fix for capybara/chromedriver issue
Our feature tests using Capybara and Chromedriver are intermittently failing with the following error:

unknown error: unhandled inspector error: {"code":-32000,"message":"Node with given id does not belong to the document"}

This can usually be fixed by rerunning the tests.

In our CI environments and end-to-end tests we pin chrome to 128 where
the issue doesn't occur.

It's harder to pin chrome in local development environments.

This commit takes a different approach found in a comment on the
capabara issue: It adds a monkey patch to the Selenium driver to
intercept the initialization of `UnknownError` and raise a
`StaleElementReferenceError` instead.

This makes the test retry.

This commit can be reverted once the issue is resolved.

https://github.com/teamcapybara/capybara/issues/2800

unknown error: unhandled inspector error: {"code":-32000,"message":"Node with given id does not belong to the document"}