Strip all HTML comments from input before rendering Markdown
When running locally, where ViewComponents are used, HTML comments
were being added with the path of the component file. The Markdown
generator was converting the `--` in the comments into em dashes,
causing the comments to be rendered as text.

Strip out all HTML comments before rendering the Markdown to avoid
this issue.