Remove have_text from tests for contact page
The contact form page feature test for a member of the public is failing.

The app redirects to the contact page at https://www.gov.uk/contact.

When the tests follow the redirect, most of the time the test fails
because the browser cannot connect to the server.

I don't know why this is happening. When running the tests with
show_browser_during_tests: true, the sae thing happens. It's possible to
hit the refresh button and get the page to load without issues.

This commit removes the have_text test for the contact page. The test
checks the current URL, which is probably enough to test the redirect.

I've changed the assertion to use have_current_path instead. This does
the same thing but is more specific than using eq on the value directly.