Fix review apps
In PR #919 we stopped setting SECRET_KEY_BASE in the Dockerfile, as
docker buildx was warning about including secrets in a Docker image.

However, the review apps for this app were relying on the secret key
being set already, rather than setting the key in the environment as in
the other apps.

The lack of a secret key base was causing the review apps to fail to
start.

This commit fixes the review apps by copying the secret key base
environment variable configuration from the review apps configuration
for our other apps.