Update forthcoming features and add 'Other features' section
Updated 'What we're working on now' section to remove the JSON bullet and add a bullet for making metrics available for more than 7 days.

Added an 'Other features' H2 at the bottom of the page to include the existing sentence about knowing there are more features that'll be useful. Also added a new sentence inviting feature feedback or requests.