Add JSON schema for submissions
Add a route for fetching the JSON schema for submissions. Currently
this will only be linked to from the JSON submission files themselves.

The schema is versioned by a single number (e.g., v1) which is
specified in the URL and the title of the schema. When we make
breaking changes, we should publish the schema with a new version.