Fix npm package lockfile
There was a bug in npm v11.6.0 and prior that would mark dependencies as
peer dependencies incorrectly (see
https://github.com/npm/cli/pull/8579). This was fixed in v11.6.1 (see
the [npm changelog], but it seems that at some point our package
lockfile was updated with incorrectly marked peer dependencies, and now
when we run `npm install` on our machines with the latest version of npm
it updates the package lockfile to remove the marks, leading to noise in
git.

This commit updates the package lockfile with the (hopefully) correct
peer dependency marks.

I don't think we need to worry about Dependabot changing things back, as
looking at the logs of a [recent Dependabot run] it looks like that is
now using npm v11.6.2.

[npm changelog]: https://github.com/npm/cli/blob/latest/CHANGELOG.md#1161-2025-09-23
[recent Dependabot run]: https://github.com/alphagov/forms-runner/actions/runs/20049289324/job/57501554830