Use GovukMarkdown mixin for ApplicationMarkdown
Instead of ApplicationMarkdown inheriting from GovukMarkdown::Renderer
and reimplementing the interface expected by markdown-rails (which is
not the expected way of using that gem), make ApplicationMarkdown look
more like the template file from markdown-rails and mixin the
govuk-markdown functionality.

This makes our code much simpler, and it easier to manage updates of
the `markdown_rails` gem.